Daemonize things

+++ newmix/Pool.py
@@ -31,21 +31,21 @@ import timing
 
 class Pool():
     def __init__(self):
-        self.next_process = timing.future(mins=1)
+        self.outbound_trigger_time = timing.future(mins=1)
         log.info("Initialised pool. Path=%s, Interval=%s, Rate=%s%%, "
                  "Size=%s.",
-                 config.get('pool', 'path'),
+                 config.get('pool', 'outbound_pool'),
                  config.get('pool', 'interval'),
                  config.getint('pool', 'rate'),
                  config.getint('pool', 'size'))
         log.debug("First pool process at %s",
-                  timing.timestamp(self.next_process))
+                  timing.timestamp(self.outbound_trigger_time))
 
     def filename(self):
         """ Return a unique, fully-qualified, random filename within the pool
             folder.
         """
-        fn = os.path.join(config.get('pool', 'path'),
+        fn = os.path.join(config.get('pool', 'outbound_pool'),
                           'm' + Random.new().read(4).encode('hex'))
         while os.path.isfile(fn):
             fn = os.path.join(config.get('pool', 'path'),
@@ -58,31 +58,25 @@ class Pool():
         with open(self.filename(), 'w') as f:
             f.write(msg)
 
-    def foo(self):
-        for n in range(500):
-            f = open(self.filename(), 'w')
-            f.write("\n")
-            f.close()
-            
+    def outbound_trigger(self):
+        return timing.now() >= self.outbound_trigger_time()
 
-    def delete(self, fqfn):
-        """Delete files from the Mixmaster Pool."""
-        head, tail = os.path.split(fqfn)
-        assert head == config.get('pool', 'path')
-        os.remove(fqfn)
-        log.debug("%s: Deleted", tail)
-
-    def select(self):
+    def outbound_select(self):
         """Pick a random subset of filenames in the Pool and return them as a
         list.  If the Pool isn't sufficiently large, return an empty list.
         """
-        files = os.listdir(config.get('pool', 'path'))
+        # Set the point in the future at which another outbound pool run will
+        # occur.
+        next_outbound = timing.dhms_future(config.get('pool', 'interval'))
+        self.outbound_trigger_time = next_outbound
+        files = os.listdir(config.get('pool', 'outbound_pool'))
         size = len(files)
         log.debug("Pool contains %s messages", size)
         if size < config.getint('pool', 'size'):
             # The pool is too small to send messages.
             log.info("Pool is insufficiently populated to trigger sending.")
             files = []
+            size = 0
         process_num = (size * config.getint('pool', 'rate')) / 100
         log.debug("Attempting to send %s messages from the pool.", process_num)
         assert process_num <= size
@@ -98,23 +92,38 @@ class Pool():
             start = random.randint(0, startmax - 1)
         end = start + process_num
         for f in files[start:end]:
-            yield os.path.join(config.get('pool', 'path'), f)
+            yield os.path.join(config.get('pool', 'outbound_pool'), f)
+
+    def outbound_delete(self, fqfn):
+        """Delete files from the Mixmaster Pool."""
+        if os.path.isfile(fqfn):
+            head, tail = os.path.split(fqfn)
+            assert head == config.get('pool', 'outbound_pool')
+            os.remove(fqfn)
+            log.debug("%s: Deleted", tail)
+        else:
+            log.error("%s: File not found during outbound msg deletion",
+                      fqfn)
 
     def inbound_select(self):
         """Pick a random subset of filenames in the Pool and return them as a
         list.  If the Pool isn't sufficiently large, return an empty list.
         """
-        files = os.listdir(config.get('pool', 'msgstore'))
+        files = os.listdir(config.get('pool', 'inbound_pool'))
         log.debug("Processing %s inbound messages in store.", len(files))
         for f in files:
-            yield os.path.join(config.get('pool', 'msgstore'), f)
+            yield os.path.join(config.get('pool', 'inbound_pool'), f)
 
     def inbound_delete(self, fqfn):
         """Delete files from the Mixmaster Pool."""
-        head, tail = os.path.split(fqfn)
-        assert head == config.get('pool', 'msgstore')
-        os.remove(fqfn)
-        log.debug("%s: Deleted", tail)
+        if os.path.isfile(fqfn):
+            head, tail = os.path.split(fqfn)
+            assert head == config.get('pool', 'inbound_pool')
+            os.remove(fqfn)
+            log.debug("%s: Deleted", tail)
+        else:
+            log.error("%s: File not found during inbound msg deletion",
+                      fqfn)
 
 
 log = logging.getLogger("newmix.%s" % __name__)
@@ -129,6 +138,6 @@ if (__name__ == "__main__"):
     handler.setFormatter(logging.Formatter(fmt=logfmt, datefmt=datefmt))
     log.addHandler(handler)
     p = Pool()
-    generator = p.select()
+    generator = p.outbound_select()
     for f in generator:
         print f
